Refactor and delete unneccessary file
