+++ cypress/fixtures/singledata.json
@@ -1,12 +1,20 @@
 {
-    "movies": [
-    {
+    "movie": {
     "id": 436270,
+    "title": "Black Adam",
     "poster_path": "https://image.tmdb.org/t/p/original//pFlaoHTZeyNkG83vxsAJiGzfSsa.jpg",
     "backdrop_path": "https://image.tmdb.org/t/p/original//bQXAqRx2Fgc46uCVWgoPz5L5Dtr.jpg",
-    "title": "Black Adam",
-    "average_rating": 4,
-    "release_date": "2022-10-19"
+    "release_date": "2022-10-19",
+    "overview": "Nearly 5,000 years after he was bestowed with the almighty powers of the Egyptian gods—and imprisoned just as quickly—Black Adam is freed from his earthly tomb, ready to unleash his unique form of justice on the modern world.",
+    "genres": [
+    "Action",
+    "Fantasy",
+    "Science Fiction"
+    ],
+    "budget": 200000000,
+    "revenue": 384571691,
+    "runtime": 125,
+    "tagline": "The world needed a hero. It got Black Adam.",
+    "average_rating": 4
     }
-]
-}
+    }
